+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 -   `isDraft` parameter to [`FormsClient.GetFormSubmission()`](./docs/forms-client.md#getformsubmission)
+-   [`TeamMembersClient`](./docs/team-members-client.md) class
 
 ## 1.0.0